process: extract color_choice() method
